+++ src/coolqbot/plugin.py
@@ -1,7 +1,10 @@
+""" 插件相关
+"""
 import os
+import pickle
 
 from coolqbot.bot import bot
-from coolqbot.config import PLUGINS_DIR_PATH
+from coolqbot.config import DATA_DIR_PATH, PLUGINS_DIR_PATH
 
 
 class PluginManager(object):
@@ -22,3 +25,28 @@ class PluginManager(object):
                 bot.logger.error(
                     f'Import error: can not import [{plugin_name}], because {e}'
                 )
+
+
+class PluginData:
+    def __init__(self, name):
+        self._name = name
+        self._data_path = DATA_DIR_PATH / name
+        self.data = {}
+
+        self.load_pkl()
+
+    def save_pkl(self):
+        try:
+            with self._data_path.open(mode='wb') as f:
+                pickle.dump(self.data, f)
+                bot.logger.debug(f'插件{self._name}：数据保存成功')
+        except Exception as e:
+            bot.logger.error(f'插件{self._name}：数据保存失败，原因是{e}')
+
+    def load_pkl(self):
+        try:
+            with self._data_path.open(mode='rb') as f:
+                self.data = pickle.load(f)
+                bot.logger.debug(f'插件{self._name}：数据加载成功')
+        except Exception as e:
+            bot.logger.error(f'插件{self._name}：数据加载失败，原因是{e}')
